categorize methods and set them on event-handler with refactoring on listeners
