+++ src/events_handler.js
@@ -1,8 +1,63 @@
 import { projects } from './variables';
 import Project from './classes/project';
-import { renderProjects, notification, projectSelected } from './dom_handler';
+// import { notification } from './dom_handler';
 
-function toggleSavePRoject() {
+// PROJECT MANIPULATION SECTION 
+function saveProject() {
+  const name = document.getElementById('new-project-name');
+  const description = document.getElementById('new-project-description');
+
+  const project = new Project(name.value, description.value);
+  projects.push(project);
+  document.body.innerHtml += notification(`Project <strong>'${name.value}'</strong> was saved succefully`);
+  name.value = '';
+  description.value = '';
+  renderProjects();
+  toggleForm();
+  projectSelected(projects.length - 1);
+}
+
+function deleteProject() {
+  const indx = document.getElementById("desc-project").lastChild.innerHTML;
+  const name = projects[indx].name
+  if (projects.length > 1) {
+    projects.splice(indx,1);
+  } else {
+    projects[0] = new Project('My First Project', 'This is your first project, you can edit it! or add a to-do');
+  }
+  document.body.innerHtml += notification(`Project <strong>'${name}'</deleted> was deleted succefully`, 'is-danger');
+  renderProjects();
+  projectSelected(0);
+}
+// PROJECT MANIPULATION SECTION 
+
+// PROJECT MANIPULATION HELPERS
+export function projectSelected(index){
+  const project = projects[index]
+  const hide = document.createElement('span');
+  let descriptionNode = document.getElementById('desc-project');
+  hide.innerHTML = index;
+  hide.classList.add("hide");
+  
+  document.getElementById('title-project').innerHTML = project.name;
+  descriptionNode.innerHTML = project.description;
+  descriptionNode.appendChild(hide);
+  document.getElementById('delete-project').addEventListener("click", deleteProject);
+  isHighlited(index);
+}
+
+export function firstProjectSelected(){
+  projectSelected(0);
+}
+
+function describeProject(e) {
+  const index = String(e.target.id).match(/(\d+)/)[0];
+  projectSelected(index);
+}
+// PROJECT MANIPULATION HELPERS
+
+// DOM MANIPULATION
+function toggleForm() {
   const disable = document.getElementById('add-project').disabled;
   document.getElementById('form-hide').classList.toggle('hide');
   document.getElementById('add-project').disabled = !disable;
@@ -10,31 +65,51 @@ function toggleSavePRoject() {
   document.getElementById('new-project-description').value = '';
 }
 
-function saveProject() {
-  const name = document.getElementById('new-project-name');
-  const description = document.getElementById('new-project-description');
-
-  const project = new Project(name.value, description.value);
-  projects.push(project);
-  document.body.innerHtml += notification(`Project ${name.value} was saved succefully`);
-  name.value = '';
-  description.value = '';
-  renderProjects();
-  toggleSavePRoject();
-  projectSelected(projects.length - 1);
-}
-
 function dismissNotification() {
   const notContainer = document.getElementById('notification');
   notContainer.classList.add('hide');
 }
 
-export function addListenerToProjects() {
-  document.getElementById('add-project').addEventListener('click', toggleSavePRoject);
-  document.getElementById('button-save-project').addEventListener('click', saveProject);
-  document.getElementById('delete-form').addEventListener('click', toggleSavePRoject);
+function setListener() {
+  projects.forEach((elem, index) => {
+    document
+      .getElementById(`project-n-${index}`)
+      .addEventListener('click', describeProject);
+  });
 }
 
-export function addListenerToNotification() {
+export function notification(text, nClass = "is-success") {
+  const notContainer = document.getElementById('notification');
+  notContainer.classList.value = `notification ${nClass}`;
+  const note = document.getElementById('text-notification');
+  note.innerHTML = text;
+  notContainer.classList.remove('hide');
+}
+
+export function renderProjects() {
+  const listNode = document.getElementById('projects-list');
+  listNode.innerHTML = '';
+  projects.forEach((elem, index) => {
+    listNode.innerHTML += `<li id="project-n-${index}">${elem.name}</li>`;
+  });
+  setListener();
+}
+
+export function addListenerToProjects() {
+  document.getElementById('add-project').addEventListener('click', toggleForm);
+  document.getElementById('button-save-project').addEventListener('click', saveProject);
+  document.getElementById('delete-form').addEventListener('click', toggleForm);
   document.querySelector('.notification .delete').addEventListener('click', dismissNotification);
 }
+
+function isHighlited(i) {
+  projects.forEach((elem, index) => {
+    document
+      .getElementById(`project-n-${index}`)
+      .classList.remove('has-background-grey-light');
+  });
+
+  const row = document.getElementById(`project-n-${i}`);
+  row.classList.add('has-background-grey-light');
+}
+// DOM MANIPULATION
