+++ src/index.js
@@ -1,7 +1,7 @@
 import Project from './classes/project';
-import { addListenerToProjects, addListenerToNotification } from './events_handler';
+import { addListenerToProjects, addListenerToNotification, firstProjectSelected, renderProjects } from './events_handler';
 import { projects } from './variables';
-import { renderProjects, firstProjectSelected } from './dom_handler';
+// import {   } from './dom_handler';
 
 const project1 = new Project('My First Project', 'This is your first project, you can edit it! or add a to-do');
 const project2 = new Project('Office', 'Office description');
@@ -9,5 +9,4 @@ projects.push(project1, project2);
 
 renderProjects();
 addListenerToProjects();
-addListenerToNotification();
 firstProjectSelected();
